[US_UT] Remap TEMPORARY_CUSTODY values (Recidiviz/recidiviz-data#43449)
## Description of the change

We are observing a suspiciously high number of compartment sessions
where `compartment_level_1 = "INTERNAL_UNKNOWN"`. In investigating the
ones that had an inflow from incarceration, I noticed that many of the
inflowing periods had a `release_reason` that was originally mapped to
`RELEASE_FROM_TEMPORARY_CUSTODY`, but had been overwritten in
normalization. In normalization, any incarceration period with an
admission reason of `TEMPORARY_CUSTODY` gets its release reason updated
to `RELEASE_FROM_TEMPORARY_CUSTODY`. Since people are regularly
transferred to different housing units during periods of "temporary
custody" in UT, that normalization change was causing the actual release
to be obscured and overwritten to "INTERNAL_UNKNOWN".

I tested this remap of admission / release reasons to move away from the
temporary custody designation, and there were ~5000 fewer sessions where
`compartment_level_1 = "INTERNAL_UNKNOWN"` after this change, because we
can now observe the releases that were previously being obscured.

I put on the agenda for my next meeting with the UT data team to ask
about DIAGNOSTIC legal statuses specifically, and whether we should
consider those to be periods of temporary custody by design. if so, I
think we will likely want to map them back to TEMPORARY_CUSTODY, and
override the normalization process for those periods in UT only.

[sandbox ingest
](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-06-11_11_29_47-16935178974323385650;graphView=0?project=recidiviz-staging&inv=1&invt=Abz2_A&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7811bc42214236cea148d73bc33352392ed54e67