[US_NE] Fix csv write + db transaction level (Recidiviz/recidiviz-data#43402)
## Description of the change

in NE, we use `default_ignore_quotes: True` with custom delimiter and
newline character so the raw files do not escape any characters.
however, the csv module requires an escapechar to be set when using
`csv.QUOTE_NONE` and escapes all double quotes and newlines. since we
are not wanting to do any special quoting or escaping we can just do a
simple write to file using the custom delimiter and newline.

also, not all of NE's databases support snapshot isolation level, so we
need to use the default isolation level instead.

lastly, remove `# TODO(Recidiviz/recidiviz-data#41296): Also add a unittest that prevents this
flag from being set for NE` since i added that UT in an earlier PR.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#41296

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 353ab88125dd79563dcdba0e96d727fa908ec9b0