[Insights] Add outlier metric configs for parole and probation MI metrics (Recidiviz/recidiviz-data#41831)
## Description of the change

This is basically
https://github.com/Recidiviz/recidiviz-data/pull/40504, but with an
extra commit at the end to use the new `feature_variant` and
`inverse_feature_variant` attributes. Using a separate PR instead of
that one because it was easier to cherrypick than to rebase.

In all honesty we could probably get away with putting this in the same
release as https://github.com/Recidiviz/recidiviz-data/pull/42798 and
https://github.com/Recidiviz/recidiviz-data/pull/40888, but we're not
really in a rush so may as well wait and get this in the following rc.
Those were just merged, so this can wait until after the rc is cut on
June 13.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes https://github.com/Recidiviz/recidiviz-data/issues/40532

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Gina Valderrama <gina@recidiviz.org>
GitOrigin-RevId: 23a1778b0e1f99d744622bded3c51a4c88eb5b82