[US_UT] Fix in the way we calculate adjacent employment periods (Recidiviz/recidiviz-data#44209)
## Description of the change

We were not properly counting the time spent employed when we had two
adjacent employment periods. E.g. if a person got a new job and left the
old one, we would restart our time employed count, which was wrong. This
change fixes that.

I uploaded this into a sandbox and there's a slight increase in
eligibility: 2.5% on the eligible pop and 0.3 on the almost eligible
pop.

![Screenshot 2025-06-17 at 11 28
11 AM](https://github.com/user-attachments/assets/63056707-c9f8-497a-a2f6-1ada7ee6213a)
```
python -m recidiviz.tools.load_views_to_sandbox \
        --project_id recidiviz-staging\
        --sandbox_dataset_prefix hsalas \
        --state_code_filter US_UT manual \
        --dataset_ids_to_load task_eligibility_criteria_general,task_eligibility_spans_us_ut,task_eligibility\
        --update_descendants False
```

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f3642ccd6ddddeeb982d3910182327065a1491a7