[looker] exclude state + normalized state dashboards from copy script (Recidiviz/recidiviz-data#44220)
## Description of the change

the autogeneration script for state/normalized state outputs a single
dashboard instead of a prod/staging dashboard because in the future we
want a single dashboard that references a `model` variable that is set
to either `recidiviz-staging` or `recidiviz-123` depending on the looker
branch. but this doesnt work currently because we don't have multiple
projects set up. i have previously manually edited the dashboards so
they work in our existing looker world so exclude from sync so we don't
have to manually remove these changes every time we sync

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#36190

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3c7afc7933ce207847eda0a875b2c07159762eca