[US_IA] Ingest supervision conditions + add primary supervision status start date to metadata (Recidiviz/recidiviz-data#42544)
## Description of the change

This PR:
- ingests BOP supervision conditions for parole periods
- adds information about the primary supervision status to supervision
periods metadata
- **Context:** In IA, a client could be on multiple different
supervision statuses at the same time, each with their own start and end
dates. Because of this, IA has a concept of the "primary" supervision
status that is the prioritized supervision status across all open
statuses. In our supervision periods ingest, we ingest periods on the
supervision status level, so if a person has multiple active statuses at
once, they would have multiple overlapping open periods. To distinguish
and inform us about primary supervision statuses, this PR:
- adds a primary supervision status flag to supervision periods metadata
that flags whether that given period is associated with the primary
supervision status
- adds a primary supervision status start date field to supervision
periods metadata that represents the start date of the primary
supervision status applicable during that given period (even if that
specific supervision period is not associated with the primary
supervision status)
- Moves the list of valid supervision period statuses into a query
fragment since it's referenced multiple times in the view, and adds
"Federal" into the list
- adds "OWI Prison" into the list of statuses we include as part of
incarceration periods

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 732761026466b9183ece02d8a72ab704fe1242a6