[US_AR] Remove present without info sentence statuses (Recidiviz/recidiviz-data#43445)
## Description of the change

**tldr; we're closing sentence serving periods early from these
statuses**

We recently added `IMPOSED_PENDING_SERVING` statuses to the schema (see
https://github.com/Recidiviz/recidiviz-data/pull/43394)

When investigating the use of serving statuses, I came across the
`PRESENT_WITHOUT_INFO` statuses hydrated in AR for this very situation.
However, the hydration is a bit complex with some unexpected results...
for example getting imposition after an initial serving status (this is
maybe pre-trial serving?).

BUT I was curious about how this impacted serving periods, and the first
example I found had us closing a period on an open sentence. So this PR
removes these statuses entirely as it looked like we weren't explicitly
using them and had another source for serving statuses. Happy to make
TODOs for how to proceed moving forward.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 97c753598ab6c95b698db07c5abb2d634ea31f1f