[monitoring] introduce concept of `job_type` to `JobRun` (Recidiviz/recidiviz-data#44237)
## Description of the change

this PR introduces the concept of a `job_type` to `JobRun` and
`incident_type` to `AirflowAlertingIncident`. the idea behind this is
motivated by adding a new type of delegate that is going to monitoring
multiple different attributes of a single
`{conf_string}{self.dag_id}.{self.job_id}` (overall runtime in addition
to success status).

this is also motivated by thinking it would be helpful to see what the
failure type is in #pagerduty-alerts at the front of the alert like
`Failure: Runtime Threshold Exceeded: <airflow task name>` to help
distinguish these types of alerts from job failures

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Prework for Recidiviz/recidiviz-data#36769

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 168d0b9b1438dfa0d89d8a4e60df3a55624a1e5e