[Research] Remove DADS as code owners of `recidiviz/research` (Recidiviz/recidiviz-data#44384)
## Description of the change

Removed DADS from code ownership of `recidiviz/research`—now that we're
starting to migrate files over and work within this directory, it's
getting noisy to have all of DADS tagged on every PR in that directory.
We'll shift to having folks tag DADS for review themselves (when
relevant).

## Related issues

N/A

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a6f059a91218ca6696126cf716771340ec5c3c60