[US_TN] Create testing parity for v1 and v2 sentencing (Recidiviz/recidiviz-data#44182)
## Description of the change

This PR:
- migrates the v1 sentencing ingest view over to the v2 ingest view
tests
- removes the old parser test
- adds the v1 sentencing unit tests to the v2 sentencing test, allowing
us to directly compare v1 and v2 for these test cases

The main difference I see between v1 and v2 sentences is the format for
`conditions` which is JSON in v1 and keeps track of datetime information
alongside the conditions (v2 is just plain text of all conditions).

I'd like to use this PR as an opportunity to discuss what we'd like the
v2 implementation to look like. I will make issues to update our
implementation in ingest and downstream if necessary. (I'd like to avoid
making logic changes in this PR, as we're just updating fixture and test
organization)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| **Type: Non-breaking refactor** | change addresses some tech debt item
or prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 87069f98c18499127186c051047e84eed17457cb